Extract App.tsx state and effects into dedicated hooks
Pure hook extraction with zero behavior or JSX change — App.test.tsx
passes unmodified. App.tsx drops from 786 to ~430 lines; the moved
logic now lives in nine focused hooks under frontend/src/hooks/:

- useOidSession — OID bootstrap, init form, logout
- useHudVisibility — control-bar auto-hide timer + reveal rules
- useStaleSetPrompt — abandoned-match prompt (server-clock aware)
- useShareLinks — lazy share-link fetch + dialog state
- useCoachmark — first-use gesture tour lifecycle
- useOverlayLocaleSync — push operator UI language to the overlay
- useScoreActions — scoring/timeout/serve handlers + point-type picker
- useSetValueDialog — long-press exact-value dialog
- useButtonTheme — button colour/logo/font memos

Effect dependency arrays and the single-useMemo colour block (which
guards the React.memo subtrees) are preserved verbatim.
