+++ frontend/src/App.tsx
@@ -1,4 +1,4 @@
-import { useState, useEffect, useCallback, useMemo, useRef, FormEvent } from 'react';
+import { useState, useEffect, useCallback } from 'react';
 import { useI18n } from './i18n';
 import { useAppConfig } from './hooks/useAppConfig';
 import { useGameState } from './hooks/useGameState';
@@ -12,6 +12,15 @@ import { useMatchAlertHaptics } from './hooks/useMatchAlertHaptics';
 import { useScreenWakeLock } from './hooks/useScreenWakeLock';
 import { useKeyboardShortcuts } from './hooks/useKeyboardShortcuts';
 import { useAutoSetSummary } from './hooks/useAutoSetSummary';
+import { useOidSession } from './hooks/useOidSession';
+import { useHudVisibility } from './hooks/useHudVisibility';
+import { useStaleSetPrompt } from './hooks/useStaleSetPrompt';
+import { useShareLinks } from './hooks/useShareLinks';
+import { useCoachmark } from './hooks/useCoachmark';
+import { useOverlayLocaleSync } from './hooks/useOverlayLocaleSync';
+import { useScoreActions } from './hooks/useScoreActions';
+import { useSetValueDialog } from './hooks/useSetValueDialog';
+import { useButtonTheme } from './hooks/useButtonTheme';
 import InitScreen from './components/InitScreen';
 import ScoreboardView from './components/ScoreboardView';
 import ScoreboardSkeleton from './components/ScoreboardSkeleton';
@@ -19,39 +28,8 @@ import ConfigPanel from './components/ConfigPanel';
 import ConnectionStatus from './components/ConnectionStatus';
 import AppDialogs from './components/AppDialogs';
 import PointTypePicker from './components/PointTypePicker';
-import * as api from './api/client';
 import ErrorBoundary from './components/ErrorBoundary';
-import type { ScoreButtonFontStyle } from './components/ScoreButton';
-import { TEAM_A_COLOR, TEAM_B_COLOR, FONT_SCALES, DEFAULT_FONT_SCALE } from './theme';
-import { HUD_AUTO_HIDE_MS } from './constants';
-import { asColor, asString } from './utils/coerce';
-
-type Team = 1 | 2;
-
-interface DialogState {
-  open: boolean;
-  title: string;
-  initialValue: number;
-  maxValue: number;
-  team: Team | null;
-  isSet: boolean;
-}
-
-interface FontScale {
-  scale: number;
-  offset_y: number;
-}
-
-function getInitialOid(): string {
-  const params = new URLSearchParams(window.location.search);
-  const urlOid = params.get('oid') || params.get('control');
-  if (urlOid) return urlOid;
-  try {
-    return localStorage.getItem('volley_oid') || '';
-  } catch {
-    return '';
-  }
-}
+import { asString } from './utils/coerce';
 
 export default function App() {
   const { t, lang } = useI18n();
@@ -59,9 +37,10 @@ export default function App() {
   const { settings, setSetting } = useSettings();
   const { isPortrait, buttonSize, hasRoomForPersistentControls } = useOrientation();
 
-  const [oid, setOid] = useState<string>(getInitialOid);
-  const [oidInput, setOidInput] = useState<string>(oid);
   const [activeTab, setActiveTab] = useState<'scoreboard' | 'config'>('scoreboard');
+  const { oid, setOid, oidInput, setOidInput, handleInit, handleLogout } = useOidSession({
+    onLogout: useCallback(() => setActiveTab('scoreboard'), []),
+  });
   const swipeHandlers = useSwipeNavigation({
     onSwipeLeft: activeTab === 'scoreboard' ? () => setActiveTab('config') : undefined,
     // Route through history.back() so ConfigPanel's popstate listener can
@@ -69,22 +48,6 @@ export default function App() {
     onSwipeRight: activeTab === 'config' ? () => window.history.back() : undefined,
   });
   const [isFullscreen, setIsFullscreen] = useState<boolean>(!!document.fullscreenElement);
-  const [showControls, setShowControls] = useState(true);
-  const hideTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-
-  const resetHideTimer = useCallback(() => {
-    if (hideTimerRef.current) clearTimeout(hideTimerRef.current);
-    hideTimerRef.current = setTimeout(() => setShowControls(false), HUD_AUTO_HIDE_MS);
-  }, []);
-
-  const [dialog, setDialog] = useState<DialogState>({
-    open: false,
-    title: '',
-    initialValue: 0,
-    maxValue: 99,
-    team: null,
-    isSet: false,
-  });
 
   const {
     state,
@@ -97,29 +60,21 @@ export default function App() {
     refreshCustomization,
   } = useGameState(oid);
 
-  // Push the operator's UI language onto the overlay's customization so
-  // OBS-embedded overlays (whose URL is fixed in the streaming app and
-  // cannot carry ``?lang=``) follow language changes live. The ref
-  // pins per-``lang`` attempts so a failing backend doesn't retry on
-  // every parent re-render (only when the operator picks a new
-  // language). Invariant: the control WS broadcasts ``state_update``
-  // only, never ``customization_update`` — so a second operator's
-  // PUT cannot bounce this effect into a ping-pong.
-  const lastAttemptedLocaleRef = useRef<string | null>(null);
-  const customizationLocale = asString(customization?.['locale']);
+  // Persist the chosen OID and (re)create the backend session. Lives
+  // here rather than in useOidSession because ``initialize`` comes
+  // from ``useGameState(oid)``, which needs the hook's ``oid`` first.
   useEffect(() => {
-    if (!oid) return;
-    const attemptKey = oid + ':' + lang;
-    if (customizationLocale === lang) return;
-    if (lastAttemptedLocaleRef.current === attemptKey) return;
-    lastAttemptedLocaleRef.current = attemptKey;
-    api
-      .updateCustomization(oid, { locale: lang })
-      .then(() => refreshCustomization())
-      .catch((e) => {
-        console.warn('Failed to sync overlay locale:', e);
-      });
-  }, [oid, lang, customizationLocale, refreshCustomization]);
+    if (oid) {
+      try {
+        localStorage.setItem('volley_oid', oid);
+      } catch (e) {
+        console.warn('Failed to save OID:', e);
+      }
+      initialize();
+    }
+  }, [oid, initialize]);
+
+  useOverlayLocaleSync({ oid, lang, customization, refreshCustomization });
 
   const { pulse } = useHaptics();
   // Set / match / finished transitions vibrate via the shared
@@ -138,41 +93,11 @@ export default function App() {
 
   const [resetConfirmOpen, setResetConfirmOpen] = useState(false);
 
-  // Stale-set prompt: if the operator opens the control UI on a
-  // session whose current set has been live for more than an hour
-  // the marker was probably abandoned (the operator forgot to reset
-  // it). Surface a one-shot dialog asking whether to reset or keep
-  // going. The ``firedRef`` guard makes sure we only ever fire once
-  // per page load so a refusal isn't re-asked while the WS keeps
-  // streaming the same stale state.
-  const [stalePromptOpen, setStalePromptOpen] = useState(false);
-  const stalePromptFiredRef = useRef(false);
-  useEffect(() => {
-    if (stalePromptFiredRef.current) return;
-    if (!state) return;
-    if (state.match_finished) return;
-    // Threshold is configured server-side via the
-    // ``STALE_SET_THRESHOLD_MINUTES`` env var and arrives on the
-    // ``/api/v1/app-config`` response. ``0`` disables the prompt.
-    const thresholdSec = (appConfig.stale_set_threshold_minutes ?? 60) * 60;
-    if (thresholdSec <= 0) return;
-    const startedAt = state.current_set_started_at;
-    if (typeof startedAt !== 'number' || startedAt <= 0) return;
-    // Prefer the server's wall clock when the payload includes it
-    // (every fresh state response carries ``server_time``), so the
-    // stale-set threshold isn't tripped by a client whose system
-    // clock is hours off. Fall back to ``Date.now()`` on the rare
-    // legacy payload that omits the field.
-    const nowSec =
-      typeof state.server_time === 'number' && state.server_time > 0
-        ? state.server_time
-        : Date.now() / 1000;
-    const elapsed = nowSec - startedAt;
-    if (elapsed > thresholdSec) {
-      stalePromptFiredRef.current = true;
-      setStalePromptOpen(true);
-    }
-  }, [state, appConfig.stale_set_threshold_minutes]);
+  const { stalePromptOpen, setStalePromptOpen } = useStaleSetPrompt({
+    state,
+    thresholdMinutes: appConfig.stale_set_threshold_minutes,
+  });
+
   // Auto-trigger the set-summary recap on each set transition (operator
   // opt-in via the Behavior section). The recap waits ``delaySec`` so
   // the broadcast camera has time to linger on the players' reaction
@@ -184,25 +109,14 @@ export default function App() {
     durationSec: settings.autoShowSetSummaryDuration,
     setSetSummary: (v) => actions.setSetSummary(v),
   });
-  // Coachmark fires once the first authoritative state lands and the
-  // operator hasn't dismissed the tour yet. The dismissal flips
-  // ``settings.gestureTourSeen`` to ``true`` and persists across
-  // sessions; the Behavior section exposes a "Replay tour"
-  // affordance that flips it back to ``false`` to re-open this on
-  // demand without a page refresh.
-  const [coachmarkOpen, setCoachmarkOpen] = useState(false);
 
-  // Share dialog: lazy-fetched links, opened from the HUD's share
-  // button. Kept in App so the dialog renders on top of both the
-  // scoreboard and config tabs (useful when the operator opens it
-  // from either surface).
-  const [shareOpen, setShareOpen] = useState(false);
-  const [shareLinks, setShareLinks] = useState<{
-    control?: string;
-    overlay?: string;
-    preview?: string;
-    follow?: string;
-  } | null>(null);
+  const { coachmarkOpen, handleCoachmarkDismiss } = useCoachmark({
+    state,
+    gestureTourSeen: settings.gestureTourSeen,
+    setSetting,
+  });
+
+  const { shareOpen, setShareOpen, shareLinks, handleOpenShare } = useShareLinks(oid);
 
   // Recent-audit drawer: a non-modal slide-in panel that surfaces
   // the per-OID action log so the operator can verify what just
@@ -213,11 +127,6 @@ export default function App() {
   // Keyboard shortcuts help modal — opened with `?`, listed in the
   // Behavior section as a "Show shortcuts" entry.
   const [shortcutsHelpOpen, setShortcutsHelpOpen] = useState(false);
-
-  // Per-point classification picker. Holds the team whose score tap is
-  // awaiting a point-type choice, or ``null`` when closed. Only used
-  // when ``settings.trackPointTypes`` is on.
-  const [pointPickerTeam, setPointPickerTeam] = useState<Team | null>(null);
 
   // Gate preview fetch on session readiness — /api/v1/links returns 404 until
   // initSession has created the session.
@@ -241,49 +150,11 @@ export default function App() {
     compactLandscape ? 5 : 8,
   );
 
-  // Reveal the bar when the viewport gains room for it (e.g. phone→tablet
-  // resize). Kept in its own effect so the manual hide toggle still works
-  // on tablets — the auto-hide effect below would otherwise re-show it on
-  // every showControls change.
-  useEffect(() => {
-    if (hasRoomForPersistentControls) {
-      setShowControls(true);
-    }
-  }, [hasRoomForPersistentControls]);
-
-  // Reveal the bar whenever the match flips back to the pending state
-  // (initial load or post-reset) so the operator can see the Start button.
-  const matchStartedAt = state?.match_started_at ?? null;
-  useEffect(() => {
-    if (matchStartedAt == null) {
-      setShowControls(true);
-    }
-  }, [matchStartedAt, setShowControls]);
-
-  useEffect(() => {
-    // On tablets/desktops the control bar fits without covering scoreboard
-    // elements, so skip the inactivity timer entirely.
-    if (hasRoomForPersistentControls) return;
-    // Keep the bar visible while the match is pending — only arm the
-    // inactivity timer once ``match_started_at`` is stamped. Also covers
-    // the pre-init case where ``state`` itself is still null.
-    if (state?.match_started_at == null) return;
-    // When the set-summary recap is live, the operator must be able to
-    // turn it off in one tap — never auto-hide the HUD.
-    if (state?.set_summary) {
-      if (hideTimerRef.current) clearTimeout(hideTimerRef.current);
-      setShowControls(true);
-      return;
-    }
-    if (showControls && activeTab === 'scoreboard') {
-      resetHideTimer();
-      window.addEventListener('pointerdown', resetHideTimer, { passive: true });
-    }
-    return () => {
-      if (hideTimerRef.current) clearTimeout(hideTimerRef.current);
-      window.removeEventListener('pointerdown', resetHideTimer);
-    };
-  }, [showControls, activeTab, state, resetHideTimer, hasRoomForPersistentControls]);
+  const { showControls, setShowControls } = useHudVisibility({
+    hasRoomForPersistentControls,
+    activeTab,
+    state,
+  });
 
   const setsLimit = (state?.config?.sets_limit as number | undefined) ?? 5;
   const matchFinished = state?.match_finished ?? false;
@@ -306,80 +177,17 @@ export default function App() {
     }
   }, [state, computeCurrentSet]);
 
-  const handleInit = useCallback(
-    (e?: FormEvent<HTMLFormElement>) => {
-      e?.preventDefault();
-      if (oidInput.trim()) {
-        setOid(oidInput.trim());
-      }
-    },
-    [oidInput],
-  );
-
-  useEffect(() => {
-    if (oid) {
-      try {
-        localStorage.setItem('volley_oid', oid);
-      } catch (e) {
-        console.warn('Failed to save OID:', e);
-      }
-      initialize();
-    }
-  }, [oid, initialize]);
-
-  // Score a point (optionally tagged). Shared by the direct tap path
-  // and the point-type picker so the auto-simple side-effect stays in
-  // one place.
-  const commitPoint = useCallback(
-    (team: Team, pointType?: api.PointType, errorType?: api.ErrorType) => {
-      actions.addPoint(team, false, pointType, errorType);
-      if (settings.autoSimple && !simpleMode) {
-        actions.setSimpleMode(true);
-      }
-    },
-    [actions, settings.autoSimple, simpleMode],
-  );
-
-  const handleAddPoint = useCallback(
-    (team: Team) => {
-      if (matchFinished) return;
-      // Opt-in classification: defer scoring to the picker so the
-      // operator can tag how the point was won. Off by default — the
-      // tap scores immediately, unchanged.
-      if (settings.trackPointTypes) {
-        setPointPickerTeam(team);
-        return;
-      }
-      commitPoint(team);
-    },
-    [matchFinished, settings.trackPointTypes, commitPoint],
-  );
-
-  const handleAddSet = useCallback(
-    (team: Team) => {
-      if (matchFinished) return;
-      actions.addSet(team, false);
-    },
-    [actions, matchFinished],
-  );
-
-  const handleAddTimeout = useCallback(
-    (team: Team) => {
-      if (matchFinished) return;
-      actions.addTimeout(team, false);
-      if (settings.autoSimple && settings.autoSimpleOnTimeout && simpleMode) {
-        actions.setSimpleMode(false);
-      }
-    },
-    [actions, matchFinished, settings.autoSimple, settings.autoSimpleOnTimeout, simpleMode],
-  );
-
-  const handleChangeServe = useCallback(
-    (team: Team) => {
-      actions.changeServe(team);
-    },
-    [actions],
-  );
+  const {
+    commitPoint,
+    handleAddPoint,
+    handleAddSet,
+    handleAddTimeout,
+    handleChangeServe,
+    handleDoubleTapScore,
+    handleDoubleTapTimeout,
+    pointPickerTeam,
+    setPointPickerTeam,
+  } = useScoreActions({ actions, settings, simpleMode, matchFinished, pulse });
 
   const handleToggleVisibility = useCallback(() => {
     if (state) actions.setVisibility(!state.visible);
@@ -409,7 +217,7 @@ export default function App() {
   // Server-side LIFO: ``actions.undoLast()`` posts to /game/undo,
   // which pops from the audit log and reverses the action via the
   // existing per-type ``add_*(undo=True)`` path. The two undo
-  // entry points (this and the per-team double-tap below) share
+  // entry points (this and the per-team double-tap) share
   // the same stack on the server and cannot drift.
   const handleUndoLast = useCallback(() => {
     pulse('confirm');
@@ -440,63 +248,13 @@ export default function App() {
     actions.reset();
   }, [actions]);
 
-  // Open the coachmark whenever the operator has unseen-tour state
-  // and authoritative game state is available. The condition stops
-  // re-firing once dismissal flips ``gestureTourSeen`` to ``true``
-  // — on the next dep change the effect runs, the guard fails, and
-  // the open state stays as the operator left it. "Replay tour"
-  // flips ``gestureTourSeen`` back to ``false`` and the effect
-  // re-opens immediately without a page refresh.
-  useEffect(() => {
-    if (state && !settings.gestureTourSeen) {
-      setCoachmarkOpen(true);
-    }
-  }, [state, settings.gestureTourSeen]);
-
-  const handleCoachmarkDismiss = useCallback(() => {
-    setCoachmarkOpen(false);
-    setSetting('gestureTourSeen', true);
-  }, [setSetting]);
-
-  const handleOpenShare = useCallback(async () => {
-    if (!oid) return;
-    setShareOpen(true);
-    if (!shareLinks) {
-      try {
-        const links = await api.getLinks(oid);
-        setShareLinks({
-          control: typeof links?.control === 'string' ? links.control : '',
-          overlay: typeof links?.overlay === 'string' ? links.overlay : '',
-          preview: typeof links?.preview === 'string' ? links.preview : '',
-          follow: typeof links?.follow === 'string' ? links.follow : '',
-        });
-      } catch {
-        // Empty links surface as the "No links available" fallback
-        // already rendered by LinksDialog — no extra error UI needed.
-        setShareLinks({});
-      }
-    }
-  }, [oid, shareLinks]);
-
   const handleStartMatch = useCallback(() => {
     actions.startMatch();
   }, [actions]);
 
-  const handleLogout = useCallback(() => {
-    try {
-      localStorage.removeItem('volley_oid');
-    } catch (e) {
-      console.warn('Failed to remove OID:', e);
-    }
-    setOid('');
-    setOidInput('');
-    setActiveTab('scoreboard');
-  }, []);
+  const { dialog, handleLongPressScore, handleLongPressSet, handleDialogSubmit, closeDialog } =
+    useSetValueDialog({ state, currentSet, setsLimit, actions, t });
 
-  // Per-team double-tap undoes the most recent forward of the
-  // same (action, team). The server-side per-type undo path now
-  // pops the matching forward from the audit log on its own, so
-  // no client-side bookkeeping is required.
   // Keyboard shortcuts. Disabled while any dialog/coachmark is open
   // (those own focus and ESC handling) or on touch-only devices where
   // the operator opted out via ``settings.keyboardShortcuts``.
@@ -519,112 +277,8 @@ export default function App() {
     onOpenHelp: () => setShortcutsHelpOpen(true),
   });
 
-  const handleDoubleTapScore = useCallback(
-    (team: Team) => {
-      pulse('confirm');
-      actions.addPoint(team, true);
-    },
-    [actions, pulse],
-  );
-
-  const handleDoubleTapTimeout = useCallback(
-    (team: Team) => {
-      pulse('confirm');
-      actions.addTimeout(team, true);
-    },
-    [actions, pulse],
-  );
-
-  const handleLongPressScore = useCallback(
-    (team: Team) => {
-      if (!state) return;
-      const teamState = team === 1 ? state.team_1 : state.team_2;
-      const rawScore = teamState.scores?.[`set_${currentSet}`];
-      const currentScore = typeof rawScore === 'number' ? rawScore : 0;
-      setDialog({
-        open: true,
-        title: t('dialog.setScore', { team }),
-        initialValue: currentScore,
-        maxValue: 99,
-        team,
-        isSet: false,
-      });
-    },
-    [state, currentSet, t],
-  );
-
-  const handleLongPressSet = useCallback(
-    (team: Team) => {
-      if (!state) return;
-      const teamState = team === 1 ? state.team_1 : state.team_2;
-      setDialog({
-        open: true,
-        title: t('dialog.setSets', { team }),
-        initialValue: teamState.sets,
-        maxValue: Math.ceil(setsLimit / 2),
-        team,
-        isSet: true,
-      });
-    },
-    [state, setsLimit, t],
-  );
-
-  const handleDialogSubmit = useCallback(
-    (value: number) => {
-      if (dialog.team === null) return;
-      if (dialog.isSet) {
-        actions.setSets(dialog.team, value);
-      } else {
-        actions.setScore(dialog.team, currentSet, value);
-      }
-      setDialog((d) => ({ ...d, open: false }));
-    },
-    [dialog, actions, currentSet],
-  );
-
-  // Memoize the four button colours together so the strings keep
-  // referential identity across re-renders that didn't change any
-  // colour input. Without this, every WebSocket state push would
-  // hand fresh string instances to TeamPanel/CenterPanel and defeat
-  // the React.memo wrappers that guard those subtrees.
-  const { btnColorA, btnTextA, btnColorB, btnTextB } = useMemo(
-    () => ({
-      btnColorA: settings.followTeamColors
-        ? asColor(customization?.['Team 1 Color'], TEAM_A_COLOR)
-        : (settings.team1BtnColor ?? TEAM_A_COLOR),
-      btnTextA: settings.followTeamColors
-        ? asColor(customization?.['Team 1 Text Color'], '#ffffff')
-        : (settings.team1BtnText ?? '#ffffff'),
-      btnColorB: settings.followTeamColors
-        ? asColor(customization?.['Team 2 Color'], TEAM_B_COLOR)
-        : (settings.team2BtnColor ?? TEAM_B_COLOR),
-      btnTextB: settings.followTeamColors
-        ? asColor(customization?.['Team 2 Text Color'], '#ffffff')
-        : (settings.team2BtnText ?? '#ffffff'),
-    }),
-    [
-      settings.followTeamColors,
-      settings.team1BtnColor,
-      settings.team1BtnText,
-      settings.team2BtnColor,
-      settings.team2BtnText,
-      customization,
-    ],
-  );
-
-  const iconLogoA = settings.showIcon ? asString(customization?.['Team 1 Logo']) : null;
-  const iconLogoB = settings.showIcon ? asString(customization?.['Team 2 Logo']) : null;
-
-  const fontStyle = useMemo<ScoreButtonFontStyle>(() => {
-    const fontProps: FontScale = FONT_SCALES[settings.selectedFont] ?? DEFAULT_FONT_SCALE;
-    return settings.selectedFont && settings.selectedFont !== 'Default'
-      ? {
-          fontFamily: `'${settings.selectedFont}'`,
-          fontScale: fontProps.scale,
-          fontOffsetY: fontProps.offset_y,
-        }
-      : { fontFamily: undefined, fontScale: 1.0, fontOffsetY: 0.0 };
-  }, [settings.selectedFont]);
+  const { btnColorA, btnTextA, btnColorB, btnTextB, iconLogoA, iconLogoB, fontStyle } =
+    useButtonTheme({ settings, customization });
 
   if (!oid || (error && !state)) {
     return (
@@ -737,7 +391,7 @@ export default function App() {
       <AppDialogs
         dialog={dialog}
         onDialogSubmit={handleDialogSubmit}
-        onDialogClose={() => setDialog((d) => ({ ...d, open: false }))}
+        onDialogClose={closeDialog}
         resetConfirmOpen={resetConfirmOpen}
         onResetConfirm={() => {
           confirmReset();
